setup.py: fix "openerp-client" script under different prefixes.
Before (and since the Jan 3 fix), the /usr/bin/openerp-client script
would assume that client would be installed in /usr/lib/python2.x/site-packages
This broke for installations with --prefix != '/usr'

Now, put a little more logic and let it take the correct path from the
"install" command.

Note that this step is a "build_scripts" step that happens /before/ any
install commands. This means, if you call "./setup.py build ...", then
a script will be generated, which may /not/ correspond to the correct
installation paths. However, when you run the "./setup.py install.."
command, later, the script will be corrected.