actions: allow search_view to come down the wire as struct, not str
There are 2 ways we can get the search_view from the server: either
the fields_view_get() result, or the 'search_view' field of
ir.actions.act_window. For the latter one to be a scalar value, we had
been converting both cases to str() and then eval() back to the view
dictionary.

Now, allow both formats to be sent from the server, avoid str/eval pairs
and handle/log any failure of eval().