view/list.py: handle missing record data at group trees
[Bug 743890] Triggered by: Dr. Ferdinand Gassauer (CampToCamp)

In an attempt to limit the view of (not in stock) products, Dr Gassauer
had overriden the read() method of product.product. Thus returning /less/
records than the ids requested. He did not implement the same logic for
the search() method, however.
This meant that the first level of the tree would indicate parent
categories, with empty children lists. This triggered a fatal crash of the
Gtk client:
  - if the on_iter_children() returns an empty list, we seem to end up in
    a corrupt C-wise stack (pyGtk issue). "None" seems to bail out with a
    modest Gtk error (console) message.
  - on_iter_has_child() has to look at the list of children, not only the
    'has_children' flag, to avoid the expander icon at all.

All this wouldn't have happened, in the first place, without the broken
module. But we'd better let the Gtk client survive in this case, too.