win_import: Don't let exceptions crash the client [Bug 753277]
When the CSV importing encounters an exception from the server, it tries
to raise an exception dialogue, passing some "parent" window object.
We used to pass the import window as a parent, which is wrong because
that one is already destroyed before that code. This must have been the
reason Gtk crashed.