[FIX] one2many_list: prevent crash when evaluating context during init
Some weird views define context values that contain references
to builtins, which may cause random behavior and errors.
For example account.invoice sets a {'type':type} context on its
invoice lines, which may use the 'type' builtin if no record
is loaded with a real 'type' field yet.
This is normally taken care of by using record.expr_eval() instead
of tools.expr_eval, , but it is not available yet at this point