form_gtk/selection: use the index rather than string matching
Let's see if this works better: one customer has come to the border
case where 2x selection strings (in a many2one backend) are identical,
making the choice arbitrary.

It's still not fully usable, because the user will have to try several
combinations of selecting either the 1st or 2nd occurence, or even
need to select one different value in between. But enough to "unhide"
that second choice.